Divide RAG to multiple components, fix linter
